Adds new ways to die!
The switches now correspond with how the RGB LED lights up to indicate the proper keypad code.
Typing in the wrong code kills the player.

Blocking a laser will kill the player.

TODO: The updateLed function makes some assumptions as to how the switches return. A verification needs to take place as this may not work as Graigry thought.