Debugs problem with code.
The issue is that many of the commands sent by the Webserver are still sent via the old shared queue method. However, as far as I can tell, the logic isn't access it. All of these funtions need to be changed to use the shared dictionary instead. I started work on changing the start button to use the dictionary, but there are still some bugs to work out.